Remove dependency on planetlab/session.py
This change removes the dependency model.py has on planetlab/session.py
by moving class methods 'sync' from the model definitions of Site,
Node, and Slice to functions in sync.py.

The original code is severely lacking consistency with current code
quality and style and should be revised in a future change.