Add zone generation support to mlabconfig.py (#39)
* Add "zone" as a supported format for mlabconfig.

This change adds a new output format for mlabconfig to generate a zone
file for measurement-lab.org. This form of generation should obsolete
gen-zone.py.

The mlabzone.header.in is taken from the current gen-zone.py script. The
included version of mlabzone.header.in increases the zone expiration
time from 1 day to 7 days. This should allow for longer periods of
outages of the shadow master server.

* Use style-specified quotaion marks on string constants.
* Remove legacy comments from header template.
* 60 min refresh cycle / 10 min retry
* Update primary / secondary DNS servers.
* Add TODO to see if we can retire DONAR.
* Allow serial number to be provided on the command line.
* Store prefix and revision in a single tmp json file to increment revision numbers.
* Correct 4-space indenation throughout.
* Update sitename TODO to a statement.